requirements.txt
aiohttp==3.7.4.post0
aiosqlite==0.16.1
astroid==2.5.6
async-timeout==3.0.1
asyncio==3.4.3
attrs==21.2.0
autopep8==1.5.7
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
cryptography==3.4.7
cssselect==1.1.0
decorator==5.0.7
dicttoxml==1.7.4
docutils==0.17.1
future==0.18.2
geocoder==1.38.1
geoip2==4.1.0
get-mac==0.8.2
greenlet==1.1.0
idna==2.10
importlib-metadata==4.0.1
ip2geotools==0.1.5
IP2Location==8.5.1
iso8601==0.1.14
isort==5.8.0
jeepney==0.6.0
keyring==23.0.1
lazy-object-proxy==1.6.0
lxml==4.6.3
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.1.0
nest-asyncio==1.5.1
packaging==20.9
pip-review==1.1.0
pkg-resources==0.0.0
pkginfo==1.7.0
publicip==1.0.1
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
pylint==2.8.2
pyparsing==2.4.7
pypika-tortoise==0.1.0
pyquery==1.4.3
pytz==2020.5
ratelim==0.1.6
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
schedule==1.1.0
SecretStorage==3.3.1
selenium==3.141.0
six==1.16.0
speedtest-cli==2.1.3
SQLAlchemy==1.4.14
toml==0.10.2
tortoise-orm==0.17.2
tqdm==4.60.0
twine==3.4.1
typed-ast==1.4.3
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.4
webencodings==0.5.1
wrapt==1.12.1
yarl==1.6.3
zipp==3.4.1
